[Fix] Removed executing main(), and added event, context arguments in main()
When the main() function is executed as an entry point in CloudFunction,
(event, context) are automatically specified as arguments.

Because of the TypeError below,
the event and context arguments were added to the potential arguments of main().

````
TypeError: main() takes 0 positional arguments but 2 were given
```

- Document: https://cloud.google.com/functions/docs/writing/background?hl=ja#function_parameters

Since both (event,context) arguments are not used in main.py, Pylint was disabled.