[Add] #1 Github Actions for pylint with push and commits
Close #1

The current flow is to run Pylint in the development environment and commit.

This means that the execution of Pylint depends on developers,
so Pylint should be executed every time a commit is pushed by GitHub Actions.

[Fix] Delete unncessary identifier in GitHub Actions

Resolve below error in GitHub Actions.

```
The workflow is not valid. .github/workflows/pylint.yml (Line: 25, Col: 13):
The identifier 'pip-cache' may not be used more than once within the same scope.
```

https://github.com/f-teruhisa/stooqifier/actions/runs/505546024