[Fix] Just execute main() to make all the processes ready to be executed for #3
Close #3

Variables are set outside the scope of `main()`,
and the entire process cannot be executed just by executing `main()`.

This makes it impossible to run Python from the CLI(ex. `python main.py`) to complete the process,
so we will make it easier to complete the process on the command line.

Specifically, reated `Env()` class that reads environment variables,
and cut it out into an `env.py` from `main.py`.